requirements.txt
diffusers
transformers
torch
torchvision
numpy
scikit-learn
datasets
huggingface-hub
matplotlib
opencv-python
accelerate
seaborn
scikit-image
tqdm
torchmetrics
lpips
pillow
clean-fid
dotenv